move to ofxGuiExtended
This is a weird thing. Maybe doing too much on the base lol.

moving to a new library in this branch also rewrote using ofParameter. So I guess if I move back to just ofxGui at least it's in pointers now lol.

also if you're using the version of openFrameworks that is included with VW2.5 you might run into some json issues. I think my solution was to update the " lib " folder from a more current OF release.